pyproject.toml
[project]
name = "sylviebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
